renaming package boot to autoconfigure according to springboots convention
