+++ azure-application-insights-spring-boot-starter/src/main/java/com/microsoft/applicationinsights/autoconfigure/initializer/SpringBootTelemetryInitializer.java
@@ -1,4 +1,4 @@
-package com.microsoft.applicationinsights.boot.initializer;
+package com.microsoft.applicationinsights.autoconfigure.initializer;
 
 import com.microsoft.applicationinsights.extensibility.TelemetryInitializer;
 import com.microsoft.applicationinsights.extensibility.context.DeviceContext;
