+++ azure-application-insights-spring-boot-starter/src/test/java/com/microsoft/applicationinsights/autoconfigure/ApplicationInsightsTelemetryAutoConfigurationTests.java
@@ -19,7 +19,7 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.boot;
+package com.microsoft.applicationinsights.autoconfigure;
 
 import static org.assertj.core.api.Assertions.assertThat;
